move type to sta/ap modules where appropriate
